Support jumping to headings by hash in URL
